Update setTopBlock to long regarding interface

+++ src/test/java/io/smilo/commons/peer/CoreNetworkState.java
@@ -83,7 +83,7 @@ public class CoreNetworkState implements INetworkState {
         return topBlock;
     }
 
-    public void setTopBlock(int topBlock) {
+    public void setTopBlock(long topBlock) {
         this.topBlock = topBlock;
         updateCatchupMode();
     }
